Move the random DN uncertainty function to the instrument module since it's valid for standard resolution also.
Add updated imports to support move of random DN uncertainty function.
